# backend/src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Quantum Wallet Backend API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/wallet/{id}": {
      "get": {
        "summary": "Get wallet by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "wallet"
          }
        }
      }
    },
    "/api/wallet": {
      "post": {
        "summary": "Create wallet",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "user_id",
                  "type"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "created"
          }
        }
      }
    },
...
  },
...
}